requirements.txt
feedparser==6.0.10
python-dotenv==1.0.0
openai==1.12.0 